原料管理修改及零碎修改完成

+++ src/main/java/com/kwy/management/entity/MaterialOverview.java
@@ -24,7 +24,7 @@ public class MaterialOverview implements Serializable {
 
     private String category;
 
-    private Double number;
+    private Integer number;
 
     private LocalDate updateDate;
